+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Releases are recorded as git tags in the [Github releases](https://github.com/learningequality/kolibri-design-system/releases) page.
 
+## Develop (version not yet known)
+- [#351] - Wrap `KCheckbox` default slot's content in <label>
+
+
 ## Version 1.4.x
 - [#185] - Handle arrow key navigation and improve focusOutline
 - [#338] - Allow for new 'nav' slot inline in the toolbar